Fix see-through splats on mobile: highp shader floats
On Apple GPUs mediump really is fp16 (desktop drivers promote it to
fp32), so position/covariance math in the splat shaders degraded badly
on iOS: warped splats, poor surface coverage, and an x-ray look.
antimatter15's shaders use highp, which is why the same scene looks
solid there. Both vertex and fragment shaders now use highp float.

Verified the sort pipeline itself is correct on desktop with an
end-to-end browser test (synthetic .splat, opaque red splat in front of
blue; front fully occludes back), so the mobile artifact was precision,
not ordering.

Also relax adaptive-DPR: shed resolution only under 32 fps (30-60 is
fine, especially on phones) instead of 48.

+++ src/splat/GaussianSplatsHelpers.ts
@@ -36,7 +36,10 @@ const GaussianSplatMaterial = /* @__PURE__ */ shaderMaterial(
     clipSign: 1.0,
   },
   `precision highp usampler2D; // Most important: ints must be 32-bit.
-  precision mediump float;
+  // highp float is critical on mobile: Apple GPUs really evaluate mediump as
+  // fp16 (desktop promotes to fp32), which wrecks position/covariance math —
+  // splats warp and surfaces turn see-through. antimatter15 uses highp too.
+  precision highp float;
 
   // Index from the splat sorter.
   attribute uint sortedIndex;
@@ -198,7 +201,7 @@ const GaussianSplatMaterial = /* @__PURE__ */ shaderMaterial(
     #endif
   }
 `,
-  `precision mediump float;
+  `precision highp float; // fp16 exp() banding is visible on mobile; keep highp
 
   uniform vec2 viewport;
   uniform float falloffCutoff;
